Replace most int64_t with ssize_t (except sha1 calculation ones).

+++ src/model.cpp
@@ -178,9 +178,9 @@ bool Model::removeUnits(size_t index)
 {
     bool status = false;
     if (index < mPimpl->mUnits.size()) {
-        auto units = *(mPimpl->mUnits.begin() + int64_t(index));
+        auto units = *(mPimpl->mUnits.begin() + ssize_t(index));
         units->removeParent();
-        mPimpl->mUnits.erase(mPimpl->mUnits.begin() + int64_t(index));
+        mPimpl->mUnits.erase(mPimpl->mUnits.begin() + ssize_t(index));
         status = true;
     }
 
@@ -273,7 +273,7 @@ bool Model::replaceUnits(size_t index, const UnitsPtr &units)
 {
     bool status = false;
     if (removeUnits(index)) {
-        mPimpl->mUnits.insert(mPimpl->mUnits.begin() + int64_t(index), units);
+        mPimpl->mUnits.insert(mPimpl->mUnits.begin() + ssize_t(index), units);
         status = true;
     }
